acme: Change emalloc parameter from uint to size_t
The current AMD64 implementation of getcallerpc requires a pointer to an
8-byte parameter, and since size_t should be 8 bytes wide on AMD64 and
the standard malloc takes a size_t argument, I changed emalloc's n to be
of type size_t.

+++ src/cmd/acme/util.c
@@ -417,7 +417,7 @@ estrdup(char *s)
 }
 
 void*
-emalloc(uint n)
+emalloc(size_t n)
 {
 	void *p;
 
@@ -430,7 +430,7 @@ emalloc(uint n)
 }
 
 void*
-erealloc(void *p, uint n)
+erealloc(void *p, size_t n)
 {
 	p = realloc(p, n);
 	if(p == nil)
